refactor: Utils / View 폴더 분리
